Add files via upload
script to download all company data from mca. get your api key from open data gov website. update the api key in line number 10. on running the script it will download all the companies registered in a particular year for example 2014 and saves the data to 2014.json file and so on. give the year ranges in line 14 and 15